+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   Transform-testNaming. [#502](https://github.com/atomist/sdm/issues/502)
 -   **BREAKING** Remove unimplemented property observesOnly. [#501](https://github.com/atomist/sdm/issues/501)
 -   **BREAKING** No longer export `EditResult`. Use `TransformResult`
+-   `CodeTransform` functions no longer need to return `Project` or `TransformResult`
 
 ### Fixed